Move builder to top-level
