+++ environments/python3/builder/build.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 set -e
-builderDir=${GOPATH}/src/github.com/fission/fission/environments/builder/cmd
+builderDir=${GOPATH}/src/github.com/fission/fission/builder/cmd
 pushd ${builderDir}
 GOOS=linux GOARCH=386 go build -o builder .
 popd
